drop bower, add travis protractor testing base
