+++ .gitignore
@@ -1,2 +1,6 @@
-/node_modules
-/source/components
+logs
+*.log
+.grunt
+build
+node_modules
+coverage
