+++ .travis.yml
@@ -1,0 +1,10 @@
+language: node_js
+node_js:
+  - "node"
+before_script:
+  - npm install grunt-cli -g
+  - export CHROME_BIN=chromium-browser
+  - "export DISPLAY=:99.0"
+  - "sh -e /etc/init.d/xvfb start"
+  - node server &
+  - sleep 3 # give server and xvfb some time to start
