--- bower.json
@@ -1,23 +1,0 @@
-{
-  "name": "angular-datetime",
-  "homepage": "https://github.com/vokal/angular-datetime",
-  "authors": [
-    {
-      "name": "Vokal",
-      "email": "webdevelopers@vokal.io"
-    }
-  ],
-  "description": "Basic date/time pickers for Angular.js",
-  "license": "MIT",
-  "devDependencies": {
-    "angular": "1.4.x"
-  },
-  "ignore": [
-    "Gruntfile.js",
-    "package.json"
-  ],
-  "dependencies": {
-    "moment": "~2.10.6",
-    "moment-timezone": "~0.5.0"
-  }
-}
