+++ package.json
@@ -3,7 +3,7 @@
   "version": "3.0.0",
   "description": "Basic date/time pickers for Angular.js",
   "homepage": "http://github.com/vokal/angular-datetime",
-  "main": "dist/index.js",
+  "main": "source/index.js",
   "bugs": {
     "url": "http://github.com/vokal/angular-datetime/issues"
   },
@@ -18,15 +18,28 @@
     "url": "http://github.com/vokal/angular-datetime"
   },
   "scripts": {
-    "dist": "grunt dist"
+    "dist": "grunt dist",
+    "start": "node server",
+    "webdriver-update": "node ./node_modules/protractor/bin/webdriver-manager update",
+    "test": "npm run webdriver-update && grunt"
   },
   "license": "MIT",
   "devDependencies": {
+    "angular": "1.4.x",
+    "angular-mocks": "1.4.x",
+    "babel-preset-es2015": "^6.1.2",
+    "babelify": "^7.2.0",
+    "browserify-istanbul": "^0.2.1",
+    "express": "^4.13.3",
     "grunt": "^0.4.5",
+    "grunt-browserify": "^4.0.1",
+    "grunt-contrib-clean": "^0.7.0",
     "grunt-contrib-less": "^0.11.4",
-    "grunt-contrib-uglify": "^0.8.0",
-    "grunt-umd": "^2.3.3",
-    "grunt-contrib-copy": "^0.8.1"
+    "grunt-istanbul": "^0.6.1",
+    "grunt-protractor-coverage": "^0.2.15",
+    "protractor": "^2.5.1",
+    "protractor-axs": "github:vokal/protractor-axs",
+    "q": "^1.4.1"
   },
   "dependencies": {
     "moment": "^2.10.6",
