+++ source/styles/angular-date-picker.less
