+++ source/styles/angular-time-picker.less
