+++ test/co-op-pickers.js
@@ -1,0 +1,15 @@
+"use strict";
+
+describe( "Co-op Pickers", function ()
+{
+    it( "should load", function ()
+    {
+        browser.get( "/" );
+    } );
+
+    it( "should run this test", function ()
+    {
+        expect( 5 + 6 ).toBe( 11 );
+    } );
+
+} );
