+++ test/date-picker.js
@@ -1,0 +1,14 @@
+"use strict";
+
+describe( "Date Picker", function ()
+{
+    it( "should load", function ()
+    {
+        browser.get( "/" );
+    } );
+
+    it( "should run this test", function ()
+    {
+        expect( 3 + 4 ).toBe( 7 );
+    } );
+} );
