+++ test/harness/harness.html
@@ -1,0 +1,29 @@
+<!doctype html>
+
+<html lang="en-US" data-ng-app="Harness">
+
+    <head>
+
+        <meta charset="utf-8">
+        <title>Date Time Pickers Harness</title>
+        <meta name="viewport" content="initial-scale=1.0, maximum-scale=1.0, user-scalable=0, width=360">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
+
+        <link rel="stylesheet" href="/build/angular-date-picker.css">
+        <link rel="stylesheet" href="/build/angular-time-picker.css">
+
+        <base href="/">
+    </head>
+
+    <body>
+
+        <div id="harness">
+
+            <h1>Harness Test</h1>
+
+        </div>
+
+        <script src="/build/test.js" defer></script>
+    </body>
+
+</html>
