+++ test/harness/index-tz.js
@@ -1,0 +1,4 @@
+"use strict";
+
+require( "./index.js" );
+require( "moment-timezone" );
