+++ test/time-picker.js
@@ -1,0 +1,15 @@
+"use strict";
+
+describe( "Time Picker", function ()
+{
+    it( "should load", function ()
+    {
+        browser.get( "/" );
+    } );
+
+    it( "should run this test", function ()
+    {
+        expect( 1 + 2 ).toBe( 3 );
+    } );
+
+} );
